Trait simdop::Mullo

pub trait Mullo<M: Multi + ElemTwice> {
    fn mullo(&self, a: M, b: M) -> M;
}

The Mullo trait is used to specify low element-wise multiplication functionality.

Required Methods

fn mullo(&self, a: M, b: M) -> M

Multiplies the elements in a and b and stores the lower halves of the results.