Crate simdop

Modules

arch

core

Traits

Add
The Add trait is used to specify element-wise addition functionality.

Mullo
The Mullo trait is used to specify low element-wise multiplication functionality.

Set1
The Set1 trait is used to specify broadcasting functionality.

Shli
The Shli trait is used to specify immediate bit-wise left shift functionality.

Shri
The Shri trait is used to specify immediate bit-wise right shift functionality.

Type Definitions

M16

M2

M32

M4

M8